Move files in folder and fix imports
